Use --format json for graph introspect introspection output
Remove --introspection-json; --format json now puts the { "__schema": ... }
object under data.introspection_response while plain output stays SDL.

Add SWAPI-driven TDD tests, shared structural parity helper, and
introspection JSON → SDL → apollo-compiler validation round-trip tests.
