+++ CHANGELOG.md
@@ -22,9 +22,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## 🚀 Features
 
-- **Add `--introspection-json` to `graph introspect` - @smyrick**
+- **`graph introspect --format json` now returns GraphQL introspection JSON - @smyrick**
 
-  `rover graph introspect` now accepts `--introspection-json` to emit the schema as GraphQL introspection JSON (`{ "__schema": ... }`) instead of SDL, matching the legacy `apollo schema:download` format. With `--format json`, `data.introspection_response` is the introspection object rather than an SDL string.
+  `rover graph introspect` with `--format json` now puts the schema as a GraphQL introspection object (`{ "__schema": ... }`) under `data.introspection_response` instead of an SDL string. Default plain output remains SDL. This is a behavior change for existing `--format json` consumers: traverse `data.introspection_response` to get the introspection object (for example, `jq '.data.introspection_response'`). No field or value transformation is needed beyond envelope traversal.
 
 - **Add `--changelog-message` to `graph publish` and `subgraph publish` - @SharkBaitDLS PR #3398 fixes #1884 #292**
