+++ Cargo.toml
@@ -311,6 +311,7 @@ portpicker = "0.1"
 predicates = { workspace = true }
 rand_regex = "0.19"
 reqwest = { workspace = true, features = ["rustls"] }
+rover-client = { workspace = true, features = ["introspection-test-support"] }
 rstest = { workspace = true }
 sealed_test = { workspace = true }
 serial_test = { workspace = true }
