+++ crates/rover-client/Cargo.toml
@@ -8,6 +8,7 @@ version = "0.0.0"
 publish = false
 
 [features]
+introspection-test-support = []
 testing = ["dep:mockall", "rover-tower/test"]
 
 [dependencies]
