+++ crates/rover-client/src/operations/graph/introspect/mod.rs
@@ -1,10 +1,16 @@
 mod introspection_json;
+#[cfg(any(test, feature = "introspection-test-support"))]
+mod introspection_parity;
 mod runner;
 mod schema;
 mod service;
 mod types;
 
+#[cfg(any(test, feature = "introspection-test-support"))]
+pub use introspection_json::introspection_json_to_validated_sdl;
 pub use introspection_json::sdl_to_introspection_json;
+#[cfg(any(test, feature = "introspection-test-support"))]
+pub use introspection_parity::assert_structural_parity;
 pub use runner::run;
 pub use schema::Schema;
 pub use service::{GraphIntrospect, GraphIntrospectError, GraphIntrospectLegacy};
